requirements.txt
selenium>=4.0.0
python-dotenv>=0.21.0
instagrapi>=1.18.0

